+++ ui/CHANGES.md
@@ -1,4 +1,5 @@
 #### Version: 1.1.5-SNAPSHOT (LibGDX 1.9.3)
+- **Added**: `Tooltip#getTarget`
 - **Changed**: It's now impossible to create `FileTypeFilter` `Rule` without providing at least one extension
 - **Changed**: `FileTypeFilter` select box won't be shown when `FileChooser` `SelectionMode` is set to `DIRECTORIES`
 - **Changed**: `FileChooser` now can be closed by pressing enter when file name field has focus